[feature #159956003] Update controller to obey SRP
